// backend/src/modules/users/users.service.d.ts
export declare class UsersService {
    static listUsers(organizationId: string): Promise<any>;
    static inviteUser(organizationId: string, data: any): Promise<any>;
    static updateUser(organizationId: string, userId: string, data: any): Promise<any>;
    static deactivateUser(organizationId: string, userId: string): Promise<any>;
}
//# sourceMappingURL=users.service.d.ts.map